Change MIT license to CC0 1.0 Universal
Since I'm ending up copying descriptions and adjusting them, I can
really claim no copyright on any of this. I'm trying to not be a monkey
that copies and pastes everything blindly, but still it's too similar.

This seems to be the most fair way to do it, but IANAL, and feedback is
welcome.